requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
billiard==3.6.4.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.31.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
Django==4.1.4
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-import-export==3.1.0
django-phonenumber-field==7.0.1
django-ratelimit==4.0.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
et-xmlfile==1.1.0
exponent-server-sdk==2.0.0
filelock==3.10.0
flower==1.2.0
frozenlist==1.3.3
google-auth==2.16.0
googlemaps==4.10.0
humanize==4.6.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupPy==1.14
MarkupSafe==2.1.1
multidict==6.0.4
odfpy==1.4.1
openai==0.27.2
openpyxl==3.1.2
packaging==22.0
pathlib==1.0.1
paystackapi==2.1.0
phonenumbers==8.13.3
Pillow==9.3.0
platformdirs==3.1.1
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
PyJWT==2.6.0
pyotp==2.8.0
python-dotenv==0.21.0
pytz==2022.7
PyYAML==6.0
redis==4.5.1
requests==2.28.1
rollbar==0.16.3
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sentry-sdk==1.23.1
six==1.16.0
sqlparse==0.4.3
tablib==3.4.0
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
virtualenv==20.21.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.2
